src/main/java/com/sample/model/Account.java
package com.sample.model;

/**
 * Created by RAJESH on 11/17/2018.
 */
public class Account {

    private int accountNumber;
    private String name;
    private int balance;

    public Account() {

    }

    public Account(int accountNumber) {
        this.accountNumber = accountNumber;
    }

    public int getBalance() {
        return balance;
    }

    public void setBalance(int balance) {
        this.balance = balance;
    }

    public int getAccountNumber() {
        return accountNumber;
    }

    public void setAccountNumber(int accountNumber) {
        this.accountNumber = accountNumber;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    @Override
    public String toString() {
        return "Account{" +
                "accountNumber=" + accountNumber +
                '}';
    }
}
